Convert Starfish to subdis project
Since starfish is composed of multiple projects, I've converted the main
starfish project into a subdirs to support this.
What was previously starfish has been moved into the harbour-starfish
directory.

Additionally as part of this change I've renamed the new subproject from
starfish to harbour-starfish in order to meet Jollas requirement for all
applications that are submittted to the offical store be prefixed with
harbour-.
https://harbour.jolla.com/faq#1.1.0