Fix layout issue in WatchSelectionPage
In one of my last few commits I managed to introduce a defect
in the layout of the watch selection page.

This defect was only visible when a watch had not been connected
to in the system settings.
Unfortunatly I was not able to get it looking !00% the same as it was,
but its at least usable now.